requirements.txt
streamlit==1.23.1
streamlit-folium==0.12.0
plotly==5.14.1
pandas==1.5.3
numpy==1.23.1
folium==0.14.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
haversine==2.8.0
Pillow==9.4.0